support rollbacks
This is some pretty awful code, but it SEEMS to work okay. There are a
few problems with it as it stands, though:

1. It will only check the DB version if the image name changes. So, if
using the "latest" tag, it won't get updated even if the image is
re-pulled.

2. Currently, the usage of ServiceAccounts is super hacky. It's just
using the default service account, and it relies on some roles/role
bindings being created externally

3. The script to fetch the DB version requires internet connectivity
currently. It's not hard to fix - it's only needed for fetching curl

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>