recreate job if image changes
There was a bug where, if the Concourse image changed before the
FetchDBVersion Job completed, the Job would continue to use the old
image. This could result in the Job never completing (if the original
image was invalid, it would remain in a ImagePullBackoff state forever).

Note that the same does not apply to the MigrateDB Job, since it is
based off of the Status' image, which can't change until the MigrateDB
Job completes.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>